refactor(usage-history): mueve modules/UsageHistory a features/usage-history
