+++ src/actions/board/index.ts
@@ -2,5 +2,3 @@ export * from './getBoardById'
 export * from './updateBoardName'
 export * from './getReminders'
 export * from './saveReminders'
-export * from './getUsageHistory'
-export * from './saveUsageHistory'
