+++ src/features/usage-history/api/actions/saveUsageHistory.ts
@@ -1,8 +1,8 @@
 'use server'
 
 import { prisma } from '@/shared/lib/prisma'
-import type { UsageHistory } from '@/modules/UsageHistory/model/usageHistory'
-import { migrateUsageHistory } from '@/modules/UsageHistory/model/usageHistory'
+import type { UsageHistory } from '../../model/usageHistory'
+import { migrateUsageHistory } from '../../model/usageHistory'
 import { requireBoardAccess } from '@/shared/lib/serverAuth'
 
 export async function saveUsageHistory({
