+++ src/features/usage-history/utils/sumUsageDurations.ts
