+++ src/shared/ui/organisms/Header.tsx
@@ -26,7 +26,7 @@ import { LogInAndLogOutMenuItem } from '@/shared/preferences/language'
 import { useSession } from '@/auth/hooks/useSession'
 import { Notes } from '@/features/notes'
 import { useTheme } from '@/shared/hooks/useTheme'
-import { useLastDurationPeriod } from '@/modules/UsageHistory/hooks/useLastDurationPeriod'
+import { useLastDurationPeriod } from '@/features/usage-history'
 import { useTypeOfView } from '@/shared/preferences/view-mode'
 import { TransitionLink } from '../atoms/TransitionLink'
 import { useBoardId } from '@/auth/state/store'
